requirements.txt
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
gevent==1.2.2
greenlet==0.4.12
grequests==0.3.0
idna==2.6
python-telegram-bot==8.1.1
requests==2.18.4
urllib3==1.22
